backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.1.12
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
idna==3.4
pymongo==4.5.0
python-dotenv~=1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.6
numpy==1.26.1
pandas==2.1.1
base58==2.1.1
djangorestframework-simplejwt==5.3.0